[Refactor] STAMPIT-121 - MissionListViewModel UIKit 제거, Snapshot 로직 뷰컨트롤러로 이동
